Resample data and compute conditional age-at-length

This function resamples cpue and age specimen data, adds ageing error and growth variability, and then computes conditional age-at-length.

srvy_comps_caal(
  specimen_data,
  cpue_data,
  r_t,
  yrs = NULL,
  bin = 1,
  boot_hauls = FALSE,
  boot_ages = FALSE,
  al_var = FALSE,
  al_var_ann = FALSE,
  age_err = FALSE,
  age_samples = NULL
)

Arguments

specimen_data

age-length specimen input dataframe

cpue_data

catch-per-unit effort input dataframe

r_t

age reader-tester input dataframe

yrs

any year filter >= (default = NULL)

bin

bin size (default = 1 cm), also can use custom length bins through defining vector of upper length bin limits, i.e., c(5, 10, 20, 35, 60), plus length group will automatically be populated and denoted with the largest defined bin + 1 (i.e., 61 for provided example)

boot_hauls

Boolean. Resample hauls w/replacement? (default = FALSE)

boot_ages

Boolean. Resample ages w/replacement? (default = FALSE)

al_var

Boolean. Include age-length variability in resampled age data? (default = FALSE)

al_var_ann

Boolean. Resample age-length variability annually or pooled across years? (default = FALSE)

age_err

Boolean. Include ageing error in resampled age data? (default = FALSE)

age_samples

If set at a value, tests reductions (and increases) in survey-level number of ages collected. To test, set at a proportion of ages collected, i.e., 0.8 or 1.2 (default = NULL)

Value

List with dataframe of conditional age-at-length (.caal).