replicate srvy_iss function for spatially-explicit input sample size for western, central and eastern gulf subregions

replicate srvy_iss function for spatially-explicit input sample size for western, central and eastern gulf subregions

srvy_iss_goa_wc_e(
  iters = 1,
  lfreq_data,
  specimen_data,
  cpue_data,
  strata_data,
  r_t,
  yrs = NULL,
  boot_hauls = FALSE,
  boot_lengths = FALSE,
  boot_ages = FALSE,
  al_var = FALSE,
  age_err = FALSE,
  region = NULL,
  save_interm = FALSE,
  match_orig = FALSE,
  save
)

Arguments

iters

number of iterations (500 recommended)

lfreq_data

input dataframe

specimen_data

input dataframe

cpue_data

input dataframe

strata_data

input dataframe

r_t

input dataframe

yrs

any year filter >= (default = NULL)

boot_hauls

resample hauls w/replacement (default = FALSE)

boot_lengths

resample lengths w/replacement (default = FALSE)

boot_ages

resample ages w/replacement (default = FALSE)

al_var

include age-length variability (default = FALSE)

age_err

include ageing error (default = FALSE)

region

region will create a folder and place results in said folder

save_interm

save the intermediate results: original comps, resampled comps (default = FALSE)

match_orig

match the computed values to gap output (default = FALSE)

save

name to save output